tender entity company id and tender code data type changed

+++ Dto/ClarificationRequestDto.cs
@@ -15,18 +15,18 @@ namespace MeroBolee.Dto
     public class ClarificationRequestResponseDto
     {
         public string Subject { get; set; }
-        public int MailId { get; set; }
+        public long MailId { get; set; }
         public DateTime SentDate { get; set; }
     }
 
     public class ClarificationResponseDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public bool IsRead { get; set; }
         public string Subject { get; set; }
         public string Body { get; set; }
         public DateTime SentOn { get; set; }
         public string SenderName { get; set; }
-        public int CorrespondenceId { get; set; }
+        public long CorrespondenceId { get; set; }
     }
 }
