+++ Dto/CorrespondenceRequestDto.cs
@@ -8,8 +8,8 @@ namespace MeroBolee.Dto
 
     public class CorrespondenceRequestDto
     {
-        public int SenderId { get; set; }
-        public int TenderId { get; set; }
+        public long SenderId { get; set; }
+        public long TenderId { get; set; }
         public string Subject { get; set; }
         public string EmailBody { get; set; }
         public DateTime EmailSentDateTime { get; set; }
@@ -18,13 +18,13 @@ namespace MeroBolee.Dto
     public class CorrespondenceRequestResponseDto
     {
         public string Subject { get; set; }
-        public int MailId { get; set; }
+        public long MailId { get; set; }
         public DateTime SentDate { get; set; }
     }
 
     public class CorrespondenceResponseDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public bool IsRead { get; set; }
         public string Subject { get; set; }
         public string Body { get; set; }
